+++ src/pages/api/notes-inside/[id]/index.ts
@@ -28,7 +28,10 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
     if (method === 'POST') {
         const { nazwa,  tekst, ostatnia_wersja } = body;
         try {
-
+            if(!user.administrator || !user.dodawanieNotatek){
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
             const result = await db.query(
                 `INSERT INTO Notatka (nazwa, osoba, tekst, ostatnia_wersja, folder) VALUES ($1, $2, $3, $4, $5) RETURNING *;`,
                 [nazwa, user.id, tekst, ostatnia_wersja, query.id]
@@ -51,6 +54,10 @@ export default async function folderHandler(req: NextApiRequest, res: NextApiRes
     } else if (method === 'DELETE') {
         // Delete a notatka
         try {
+            if(!user.administrator || !user.edytowanieCudzychNotatek){
+                res.status(401).json({ message: "Unauthorized" });
+                return;
+            }
             const result = await db.query(
                 `DELETE FROM Notatka WHERE folder = $1 RETURNING *;`,
                 [query.id]
